Library vol. 2

+++ src/main/java/com/crud/library/repository/BookDao.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Repository;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
@@ -13,12 +14,12 @@ public class BookDao {
     @PersistenceContext
     private EntityManager em;
 
-    public BigDecimal getAllAvialableCopies(Long bookId) {
+    public BigInteger getAllAvialableCopies(Long bookId) {
         String SearchQuery = "SELECT count(*) FROM LIBRARY_CRUD.COPIES c\n" +
                 "JOIN LIBRARY_CRUD.BOOKS B\n" +
                 "ON B.ID = C.BOOK_ID\n" +
                 "WHERE C.STATUS='FREE' AND B.ID= " + bookId;
-        return (BigDecimal) em.createNativeQuery(SearchQuery).getSingleResult();
+        return (BigInteger) em.createNativeQuery(SearchQuery).getSingleResult();
     }
 
     public RentalDto returnBorrowedBook(Long copiesId) {
